refactor server script code
