+++ scripts/app.js
@@ -1,286 +1,10 @@
-const { promisify } = require("util");
-const exec = promisify(require("child_process").exec);
-const fs = require("fs");
-const { MongoClient } = require("mongodb");
-
-function insertJob(payloadObj, jobTitle, jobUserName, jobUserEmail) {
-  const dbName = process.env.DB_NAME;
-  const collName = process.env.COL_NAME;
-  const username = process.env.USERNAME;
-  const secret = process.env.SECRET;
-  // create the new job document
-  const newJob = {
-    title: jobTitle,
-    user: jobUserName,
-    email: jobUserEmail,
-    status: "inQueue",
-    createdTime: new Date(),
-    startTime: null,
-    endTime: null,
-    priority: 1,
-    numFailures: 0,
-    failures: [],
-    result: null,
-    payload: payloadObj,
-    logs: {}
-  };
-
-  // we are looking for jobs in the queue with the same payload
-  // that have not yet started (startTime == null)
-  const filterDoc = { payload: payloadObj, startTime: null };
-  const updateDoc = { $setOnInsert: newJob };
-
-  const uri = `mongodb+srv://${username}:${secret}@cluster0-ylwlz.mongodb.net/test?retryWrites=true&w=majority`;
-  const client = new MongoClient(uri, { useNewUrlParser: true });
-  client.connect(err => {
-    if (err) {
-      console.log("error connecting to Mongo");
-      return false;
-    }
-    const collection = client.db(dbName).collection(collName);
-    collection.updateOne(filterDoc, updateDoc, { upsert: true }).then(
-      result => {
-        if (result.upsertedId) {
-          console.log(
-            "You successfully enqued a staging job to docs autobuilder. This is the record id: ",
-            result.upsertedId
-          );
-          return [true, result.upsertedId];
-        }
-        console.log("Already existed ", newJob);
-        return "Already Existed";
-      },
-      error => {
-        console.log(
-          "There was an error enqueing a staging job to docs autobuilder. Here is the error: ",
-          error
-        );
-        return [false, error];
-      }
-    );
-    client.close();
-  });
-}
-
-function createPayload(
-  repoNameArg,
-  branchNameArg,
-  repoOwnerArg,
-  urlArg,
-  patchArg,
-  buildSizeArg,
-  lastCommit
-) {
-  const payload = {
-    jobType: "githubPush",
-    source: "github",
-    action: "push",
-    repoName: repoNameArg,
-    branchName: branchNameArg,
-    isFork: true,
-    private: false,
-    isXlarge: false,
-    repoOwner: repoOwnerArg,
-    url: urlArg,
-    newHead: lastCommit,
-    buildSize: buildSizeArg,
-    patch: patchArg
-  };
-
-  return payload;
-}
-
-async function getBranchName() {
-  return new Promise((resolve, reject) => {
-    exec("git rev-parse --abbrev-ref HEAD", (error, stdout) => {
-      if (error !== null) {
-        console.log(`exec error: ${error}`);
-        reject(error);
-      }
-      resolve(stdout.replace("\n", ""));
-    });
-  });
-}
-
-// extract repo name from url
-function getRepoName(url) {
-  let repoName = url.split("/");
-  repoName = repoName[repoName.length - 1];
-  repoName = repoName.replace(".git", "");
-  repoName = repoName.replace("\n", "");
-  return repoName;
-}
-
-// delete patch file
-async function deletePatchFile() {
-  return new Promise((resolve, reject) => {
-    exec("rm myPatch.patch", error => {
-      if (error !== null) {
-        console.log("exec error deleting patch file: ", error);
-        reject(error);
-      }
-      resolve("successfully removed patch file");
-    });
-  });
-}
-async function getRepoInfo() {
-  return new Promise((resolve, reject) => {
-    exec("git config --get remote.origin.url", (error, stdout) => {
-      if (error !== null) {
-        console.log(`exec error: ${error}`);
-        reject(error);
-      }
-
-      const repoUrl = stdout.replace("\n", "");
-      resolve(repoUrl);
-    });
-  });
-}
-
-async function getGitEmail() {
-  return new Promise((resolve, reject) => {
-    exec("git config --global user.email", (error, stdout) => {
-      if (error !== null) {
-        console.log(`exec error: ${error}`);
-        reject(error);
-      } else {
-        resolve(stdout.replace("\n", ""));
-      }
-    });
-  });
-}
-
-async function getGitUser() {
-  return new Promise((resolve, reject) => {
-    exec("git config --global user.name", (error, stdout) => {
-      if (error !== null) {
-        console.log(`exec error: ${error}`);
-        reject(error);
-      } else {
-        resolve(stdout.replace("\n", ""));
-      }
-    });
-  });
-}
-
-async function getGitCommits() {
-  return new Promise((resolve, reject) => {
-    exec("git cherry", (error, stdout) => {
-      if (error !== null) {
-        console.log(`exec error: ${error}`);
-        reject(error);
-      } else {
-        const cleanedup = stdout.replace(/\+ /g, "");
-        const commitarray = cleanedup.split(/\r\n|\r|\n/);
-        commitarray.pop(); // remove the last, dummy element that results from splitting on newline
-        if (commitarray.length === 0) {
-          console.log(
-            "You have tried to create a staging job from local commits but you have no committed work. Please make commits and then try again"
-          );
-          reject();
-        }
-        if (commitarray.length === 1) {
-          const firstCommit = commitarray[0];
-          const lastCommit = null;
-          resolve({ firstCommit, lastCommit });
-        } else {
-          const firstCommit = commitarray[0];
-          const lastCommit = commitarray[commitarray.length - 1];
-          resolve({ firstCommit, lastCommit });
-        }
-      }
-    });
-  });
-}
-async function getGitPatchFromLocal() {
-  return new Promise((resolve, reject) => {
-    exec(`git diff origin/test2 --ignore-submodules > myPatch.patch`, error => {
-      if (error !== null) {
-        console.log("error generating patch: ", error);
-        reject(error);
-      } else {
-        fs.readFile("myPatch.patch", "utf8", (err, data) => {
-          if (err) {
-            console.log("error reading patch file: ", err);
-            reject(err);
-          }
-          resolve(data);
-        });
-      }
-    });
-  });
-}
-async function getGitPatchFromCommits(firstCommit, lastCommit) {
-  //need to delete patch file?
-  return new Promise((resolve, reject) => {
-    if (lastCommit === null) {
-      const patchCommand = "git show HEAD > myPatch.patch";
-      exec(patchCommand, error => {
-        if (error !== null) {
-          console.log("error generating patch: ", error);
-          reject(error);
-        } else {
-          fs.readFile("myPatch.patch", "utf8", (err, data) => {
-            if (err) {
-              console.log("error reading patch file", err);
-              reject(err);
-            }
-            resolve(data);
-          });
-        }
-      });
-    } else {
-      const patchCommand = `git diff ${firstCommit}^...${lastCommit} > myPatch.patch`;
-      exec(patchCommand, error => {
-        if (error !== null) {
-          console.log("error generating patch: ", error);
-          reject(error);
-        } else {
-          fs.readFile("myPatch.patch", "utf8", (err, data) => {
-            if (err) {
-              console.log("error reading patch file ", err);
-              reject(err);
-            }
-            resolve(data);
-          });
-        }
-      });
-    }
-  });
-}
-
-function validateConfiguration() {
-  const missingConfigs = [];
-
-  if (process.env.DB_NAME === undefined || process.env.DB_NAME === "") {
-    missingConfigs.push("DB_NAME");
-  }
-  if (process.env.COL_NAME === undefined || process.env.COL_NAME === "") {
-    missingConfigs.push("COL_NAME");
-  }
-  if (process.env.USERNAME === undefined || process.env.USERNAME === "") {
-    missingConfigs.push("USERNAME");
-  }
-  if (process.env.SECRET === undefined || process.env.SECRET === "") {
-    missingConfigs.push("SECRET");
-  }
-  if (missingConfigs.length !== 0) {
-    console.log(missingConfigs)
-    console.log(
-      `The ~/.config/.snootyenv file is found but does not contain the following required fields: ${missingConfigs.toString()}`
-    );
-    process.exit();
-  }
-
-  //is this good? or lazy?
-  return true
-}
+const StagingUtils = require('stagingUtils');
 
 async function main() {
   const patchFlag = process.argv[2];
   const buildSize = process.argv[3];
 
-  // validateConfiguration();
+  validateConfiguration();
 
   if (patchFlag === undefined) {
     console.log(
@@ -307,17 +31,18 @@ async function main() {
     return;
   }
 
-  const userName = await getGitUser();
-  const userEmail = await getGitEmail();
-  const url = await getRepoInfo();
-  const repoName = getRepoName(url);
-  const branchName = await getBranchName();
-  const newHead = "genericscsss";
+
+  const userName = await StagingUtils.getGitUser();
+  const userEmail = await StagingUtils.getGitEmail();
+  const url = await StagingUtils.getRepoInfo();
+  const repoName = StagingUtils.getRepoName(url);
+  const branchName = await StagingUtils.getBranchName();
+  const newHead = "newHead";
   // toggle btwn create patch from commits or what you have saved locally
   if (patchFlag === "commit") {
-    const { firstCommit, lastCommit } = await getGitCommits();
-    const patch = await getGitPatchFromCommits(firstCommit, lastCommit);
-    const payLoad = await createPayload(
+    const { firstCommit, lastCommit } = await StagingUtils.getGitCommits();
+    const patch = await StagingUtils.getGitPatchFromCommits(firstCommit, lastCommit);
+    const payLoad = await StagingUtils.createPayload(
       repoName,
       branchName,
       userName,
@@ -326,7 +51,7 @@ async function main() {
       buildSize,
       newHead
     );
-    const success = insertJob(
+    const success = StagingUtils.insertJob(
       payLoad,
       `Github Push: ${userName}/repoName`,
       userName,
@@ -336,8 +61,8 @@ async function main() {
   }
 
   if (patchFlag === "local") {
-    const patch = await getGitPatchFromLocal(branchName);
-    const payLoad = await createPayload(
+    const patch = await StagingUtils.getGitPatchFromLocal(branchName);
+    const payLoad = await StagingUtils.createPayload(
       repoName,
       branchName,
       userName,
@@ -346,7 +71,7 @@ async function main() {
       buildSize,
       newHead
     );
-    const success = insertJob(
+    const success = StagingUtils.insertJob(
       payLoad,
       `Github Push: ${userName}/repoName`,
       userName,
@@ -355,7 +80,7 @@ async function main() {
     return success;
   }
 
-  await deletePatchFile();
+  await StagingUtils.deletePatchFile();
 }
 
 main();
